# Conflicts:
#	Lymytz_Web/src/java/yvs/util/Managed.java

+++ Lymytz_Web/src/java/yvs/base/produits/ManagedArticles.java
@@ -6557,8 +6557,12 @@ public class ManagedArticles extends Managed<Articles, YvsBaseArticles> implemen
             if (y != null ? y.getId() > 0 : false) {
                 y.setDateUpdate(new Date());
                 y.setActif(!y.getActif());
+                y.setAuthor(currentUser);
+                y.setDateUpdate(new Date());
                 dao.update(y);
                 y.getArticle().setActif(y.getActif());
+                y.getArticle().setAuthor(currentUser);
+                y.getArticle().setDateUpdate(new Date());
                 dao.update(y.getArticle());
                 int index = pointsVentes.indexOf(y);
                 if (index > -1) {
@@ -6573,8 +6577,13 @@ public class ManagedArticles extends Managed<Articles, YvsBaseArticles> implemen
 
     public void removeConditionnementPoint(YvsBaseConditionnementPoint y) {
         try {
+            if (!autoriser("base_article_update_puv")) {
+                openNotAcces();
+                return;
+            }
             if (y != null ? y.getId() > 0 : false) {
                 int index = pointsVentes.indexOf(y);
+                y.setAuthor(currentUser);
                 y.setDateUpdate(new Date());
                 dao.delete(y);
                 y.setId(YvsBaseConditionnementPoint.ids--);
@@ -6590,9 +6599,12 @@ public class ManagedArticles extends Managed<Articles, YvsBaseArticles> implemen
 
     public void addConditionnementPoint(YvsBaseConditionnementPoint y) {
         try {
+            if (!autoriser("base_article_update_puv")) {
+                openNotAcces();
+                return;
+            }
             if (y != null ? y.getId() < 1 : true) {
                 int index = pointsVentes.indexOf(y);
-                y.setDateUpdate(new Date());
                 if (y.getArticle() == null) {
                     getErrorMessage("Action impossible!!!");
                     return;
@@ -6600,6 +6612,7 @@ public class ManagedArticles extends Managed<Articles, YvsBaseArticles> implemen
                 if (y.getArticle() != null ? y.getArticle().getId() < 1 : true) {
                     y.getArticle().setId(null);
                     y.getArticle().setDateUpdate(new Date());
+                    y.getArticle().setAuthor(currentUser);
                     y.setArticle((YvsBaseArticlePoint) dao.save1(y.getArticle()));
                 }
                 if (!y.getArticle().getActif()) {
@@ -6607,6 +6620,8 @@ public class ManagedArticles extends Managed<Articles, YvsBaseArticles> implemen
                     dao.update(y.getArticle());
                 }
                 y.setId(null);
+                y.setAuthor(currentUser);
+                y.setDateUpdate(new Date());
                 y = (YvsBaseConditionnementPoint) dao.save1(y);
                 if (index > -1) {
                     pointsVentes.set(index, y);
@@ -6620,6 +6635,10 @@ public class ManagedArticles extends Managed<Articles, YvsBaseArticles> implemen
 
     public void updateConditionnementPoint(RowEditEvent ev) {
         if (ev != null) {
+            if (!autoriser("base_article_update_puv")) {
+                openNotAcces();
+                return;
+            }
             YvsBaseConditionnementPoint y = (YvsBaseConditionnementPoint) ev.getObject();
             y.setDateUpdate(new Date());
             dao.update(y);
